Added placeholder unit tests to be filled in later, and separate directories for src/test
