+++ src/modules/twitchirc/connectionInfo.py
