+++ src/modules/twitchirc/connectionSettings.txt
